Revert "move quote_proptest to its own macros module"
This reverts commit bd26ba1b87335c83607c165f195a5788124f70fb.
